refactor(idea_service): extract lifecycle ops + meta reads (#163)
  before: 870
  after:  574 (-296)
